% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{couple_lookup}
\alias{couple_lookup}
\title{Dataset of couples on \emph{Strictly Come Dancing}}
\format{
\subsection{\code{couple_lookup}}{

A data frame with 330 rows and 4 columns:
\describe{
\item{series_num}{Series number}
\item{couple_name}{
Name of the couple as used on \emph{Strictly}. This will usually be the
first name of the celebrity, then "&", then the first name of the
professional dancer.
}
\item{celebrity}{Name of celebrity}
\item{professional_partner}{Name of professional dancer}
}
}
}
\source{
Wikipedia contributors, \href{https://en.wikipedia.org/wiki/Strictly_Come_Dancing_series_23}{Strictly Come Dancing series 23}, and
articles for earlier series. Used under the \href{https://creativecommons.org/licenses/by-sa/4.0/deed.en}{Creative Commons Attribution-ShareAlike 4.0 International licence}.
}
\usage{
couple_lookup
}
\description{
A dataset of celebrities and their professional partners on the BBC TV series
\emph{Strictly Come Dancing}.
}
\keyword{datasets}
