% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dance_offs}
\alias{dance_offs}
\title{Dataset of votes in \emph{Strictly Come Dancing}'s dance-offs}
\format{
\subsection{\code{dance_offs}}{

A data frame with 714 rows and 6 columns:
\describe{
\item{series_num}{Series number}
\item{week_num}{Week number}
\item{judge}{Judge name}
\item{vote}{
Couple that the judge in question voted to \emph{save} in the dance-off
}
\item{actually_voted_flag}{
Logical: \code{TRUE} if the judge in question cast a vote before the result
was known (i.e. up to the point where a majority of judges had voted for
one couple), and \code{FALSE} if the judge in question was asked once the
result was announced how they would have voted.
}
\item{casting_vote}{
Logical: indicates whether or not the judge held that week's casting
vote. Until Series 23, this was the same judge every week, who was termed
the 'head judge'. From Series 23 onwards, this role was allocated to a
different judge each week.
}
}
}
}
\source{
Wikipedia contributors, \href{https://en.wikipedia.org/wiki/Strictly_Come_Dancing_series_23}{Strictly Come Dancing series 23}, and
articles for earlier series. Used under the \href{https://creativecommons.org/licenses/by-sa/4.0/deed.en}{Creative Commons Attribution-ShareAlike 4.0 International licence}.
}
\usage{
dance_offs
}
\description{
A dataset of the votes cast by each judge in the dance-off on the BBC TV
series \emph{Strictly Come Dancing}.
}
\keyword{datasets}
