% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{judge_scores}
\alias{judge_scores}
\title{\emph{Strictly Come Dancing} dataset of judges' scores}
\format{
\subsection{\code{judge_scores}}{

A data frame with 10,913 rows and 4 columns:
\describe{
\item{id}{ID number for the dance in the \code{dances} table}
\item{judge}{Judge name}
\item{score}{Score: an integer from 1 to 10}
\item{guest_judge_flag}{
Logical: \code{TRUE} for guest judges that were not considered one of the
"main" judges for the series, and \code{FALSE} otherwise
}
}
}
}
\source{
Wikipedia contributors, \href{https://en.wikipedia.org/wiki/Strictly_Come_Dancing_series_23}{Strictly Come Dancing series 23}, and
articles for earlier series. Used under the \href{https://creativecommons.org/licenses/by-sa/4.0/deed.en}{Creative Commons Attribution-ShareAlike 4.0 International licence}.
}
\usage{
judge_scores
}
\description{
A dataset of individual judges' scores on the BBC TV series \emph{Strictly
Come Dancing}.
}
\keyword{datasets}
