% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{music}
\alias{music}
\title{Dataset of songs used on \emph{Strictly Come Dancing}}
\format{
\subsection{\code{music}}{

A data frame with 2,814 rows and 4 columns:
\describe{
\item{music_id}{
ID number for the music performed, corresponding to the value of the same
name in the \code{dances} table. Note that, while most values appear only once
in this table, some appear multiple times when a dance was performed to a
medley of pieces.
}
\item{song}{Name of musical piece}
\item{artist}{
Name of original artist, in whose style the piece is typically performed
(but see \code{theme_artist})
}
\item{theme_artist}{
Logical. In some cases the "artist" is not given in the source, and
instead some other detail is provided—usually this is the name of a wider
creative work (film, musical, etc.) in which the music is featured. To
flag this, these cases are indicated by \code{TRUE} in this column (all other
rows are \code{FALSE}).
}
}
}
}
\source{
Wikipedia contributors, \href{https://en.wikipedia.org/wiki/Strictly_Come_Dancing_series_23}{Strictly Come Dancing series 23}, and
articles for earlier series. Used under the \href{https://creativecommons.org/licenses/by-sa/4.0/deed.en}{Creative Commons Attribution-ShareAlike 4.0 International licence}.
}
\usage{
music
}
\description{
A dataset of musical pieces as used on the BBC TV series \emph{Strictly Come
Dancing}.
}
\keyword{datasets}
