% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{series_judges_lookup}
\alias{series_judges_lookup}
\title{Dataset of \emph{Strictly Come Dancing} judges, by series}
\format{
\subsection{\code{series_judges_lookup}}{

A data frame with 91 rows and 4 columns:
\describe{
\item{series_num}{Series number}
\item{judge}{
One of the "main" judges for that series. Judges are shown in separate
rows. Counts neither guest judges, nor "mentors" who sat on the panel but
did not award scores.
}
\item{head_judge_flag}{
Logical: was that judge the head judge for the series? \code{FALSE} for all
judges if there was no head judge.
}
\item{n_judges}{
Number of judges awarding scores that week. Does not count "mentors" who
sat on the panel but did not award scores.
}
}
}
}
\source{
Wikipedia contributors, \href{https://en.wikipedia.org/wiki/Strictly_Come_Dancing_series_23}{Strictly Come Dancing series 23}, and
articles for earlier series. Used under the \href{https://creativecommons.org/licenses/by-sa/4.0/deed.en}{Creative Commons Attribution-ShareAlike 4.0 International licence}.
}
\usage{
series_judges_lookup
}
\description{
A dataset of which judges were the "main" judges for each series of the BBC
TV series \emph{Strictly Come Dancing}.
}
\keyword{datasets}
