% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weekly_judges_lookup}
\alias{weekly_judges_lookup}
\title{Dataset of \emph{Strictly Come Dancing} judges}
\format{
\subsection{\code{weekly_judges_lookup}}{

A data frame with 279 rows and 4 columns:
\describe{
\item{series_num}{Series number}
\item{week_num}{Week number}
\item{weekly_judges}{
A list column. Each row's entry in the list is a factor, giving the names
of the relevant judges on the panel. Does not count "mentors" who sat on
the panel but did not award scores.
}
\item{n_judges}{
Number of judges awarding scores that week. Does not count "mentors" who
sat on the panel but did not award scores.
}
}
}
}
\source{
Wikipedia contributors, \href{https://en.wikipedia.org/wiki/Strictly_Come_Dancing_series_23}{Strictly Come Dancing series 23}, and
articles for earlier series. Used under the \href{https://creativecommons.org/licenses/by-sa/4.0/deed.en}{Creative Commons Attribution-ShareAlike 4.0 International licence}.
}
\usage{
weekly_judges_lookup
}
\description{
A dataset of which judges awarded scores in each week, for BBC TV series
\emph{Strictly Come Dancing}.
}
\keyword{datasets}
